% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chapuy_features}
\alias{chapuy_features}
\title{Features for DLBCL grouping by Chapuy method.}
\format{
\subsection{\code{chapuy_features}}{

A named list.
}
}
\usage{
chapuy_features
}
\description{
A named list of data frames and lists needed to assemble matrix and classify
DLBCL tumors according to Chapuy method described in PMID: 29713087.
}
\keyword{datasets}
