% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{classify_bl}
\alias{classify_bl}
\title{Classify BL samples into genetic subgroups.}
\usage{
classify_bl(
  these_samples_metadata,
  maf_data,
  projection = "grch37",
  output = "both",
  ashm_cutoff = 3
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id
column with ids for the samples to be classified. Required input.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. Any maf
columns can be provided, but the required are "Hugo_Symbol",
"NCBI_Build", "Chromosome", "Start_Position", "End_Position",
"Variant_Classification", "HGVSp_Short", and "Tumor_Sample_Barcode".
Required input.}

\item{projection}{The projection of the samples. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one
of predictions, matrix, or both. Defaults to both.}

\item{ashm_cutoff}{Numeric value indicating number of mutations for
binarizing aSHM feature. Recommended to use the default value (3).}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Assemble the binary feature matrix and use the random forest prediction model
to classify BL tumors into genetic subgroups. Please see PMID 36201743 on
genetic subgroups of BL.
}
\examples{
test_meta <- get_gambl_metadata()  \%>\%
    filter(pathology == "BL")
maf <- get_ssm_by_samples(
    these_samples_metadata = test_meta
)
predictions <- classify_bl(
     these_samples_metadata = test_meta,
     maf_data = maf
)
predictions <- classify_bl(
     these_samples_metadata = test_meta,
     maf_data = maf,
     output = "predictions"
)

}
