% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{classify_fl}
\alias{classify_fl}
\title{Classify FL samples into cFL/dFL subgroups.}
\usage{
classify_fl(
  these_samples_metadata,
  maf_data,
  matrix,
  output = "predictions",
  model = RFmodel_FL
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id
column with ids for the samples to be classified. Required argument.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least
must contain the first 45 columns of standard MAF format. Required
argument. The maf data must be in the grch37 projection.}

\item{matrix}{Optionally, if the binary feature matrix is already prepared,
it can be provided in this argument.}

\item{output}{The output to be returned after prediction is done. Can be one
of predictions, matrix, or both. Defaults to predictions.}

\item{model}{The RF model. Classifier from the paper describing cFL is used.
It is not recommended to change the value of this parameter. This is a
developer-only option and used for testing and functionality improvement.}
}
\value{
data frame, binary matrix, or both
}
\description{
Use the random forest prediction model to assemble the binary matrix and use
it to classify FL tummors into cFL/dFL
}
\examples{
meta <- GAMBLR.data::sample_data$meta \%>\%
    filter(pathology == "FL")

maf <- GAMBLR.data::get_coding_ssm(
    these_samples_metadata = meta
)

classify_fl(
    these_samples_metadata = meta,
    maf_data = maf
)
classify_fl(
    these_samples_metadata = meta,
    maf_data = maf,
    output = "both"
)

}
