% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{depr_nearest_neighbor_heatmap}
\alias{depr_nearest_neighbor_heatmap}
\title{Heatmap visualization of mutations in nearest neighbors for a sample}
\usage{
depr_nearest_neighbor_heatmap(
  this_sample_id,
  DLBCLone_model,
  truth_column = "lymphgen",
  metadata_cols = NULL,
  clustering_distance = "binary",
  font_size = 14
)
}
\arguments{
\item{this_sample_id}{Character. The sample ID for which to plot the nearest neighbor heatmap.}

\item{DLBCLone_model}{List. A DLBCLone model object, either from \code{DLBCLone_optimize_params}
or \code{DLBCLone_KNN} (with \code{predict_unlabeled = TRUE}).}
}
\value{
A ComplexHeatmap object showing the feature matrix for the nearest neighbors of the sample.
}
\description{
Generates a heatmap of feature values for the nearest neighbors of a specified sample,
based on a DLBCLone model object. This visualization helps to inspect the feature profiles
of samples most similar to the query sample.
}
\details{
\itemize{
\item For models of type \code{DLBCLone_optimize_params}, uses the \code{neighbors} and \code{lyseq_status} fields.
\item For models of type \code{DLBCLone_KNN}, uses the \code{unlabeled_neighbors} field.
\item The function extracts the feature matrix rows corresponding to the nearest neighbors of the specified sample,
and plots a heatmap of features with nonzero values.
}
}
\examples{
# Assuming 'predicted_out' is a the output of DLBCLone_KNN_predict
and "SomeSample_ID" is a valid sample ID from among the test (not training) samples
\dontrun{
nearest_neighbor_heatmap("SomeSample_ID", predicted_out)
}

}
