% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlbclass_features_all}
\alias{dlbclass_features_all}
\title{All DLBCLass features.}
\format{
\subsection{\code{dlbclass_features_all}}{

A data frame with 163 rows and 7 columns.
\describe{
\item{gene}{Gene symbol in Hugo format}
\item{cluster}{The associated cluster}
\item{AlterationType}{The type of alteration}
\item{Marker.Gene.in.Cluster}{TODO}
\item{Unique}{TODO}
\item{Rank.within.Cluster}{Rank of feature importance for the class}
\item{q}{q-value}
}
}
}
\usage{
dlbclass_features_all
}
\description{
The features associated with each DLBCLass class according to the Supplemental
data from Chapuy et al, 2024.
}
\keyword{internal}
