% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DLBCLone_summarize_model}
\alias{DLBCLone_summarize_model}
\title{Summarize and Export DLBCLone Model Results}
\usage{
DLBCLone_summarize_model(base_name, optimized_model)
}
\arguments{
\item{base_name}{Character. The base name (and directory) for saving output files.}

\item{optimized_model}{List. The output from DLBCLone optimization, containing predictions, features, and metadata.}
}
\value{
No return value. Side effect: writes multiple PDF files to disk.
}
\description{
Generates and saves a set of summary plots and tables for a DLBCLone model, including UMAP scatterplots, alluvial plots, and oncoplots.
Results are saved as PDF files in a directory named after the provided base name.
}
\details{
\itemize{
\item Creates a directory for results if it does not exist.
\item Saves UMAP scatterplots for all samples and for non-"Other" samples.
\item Generates alluvial plots for different DLBCLone predictions
\item Exports an oncoplot summarizing mutation and classification results.
\item Uses \code{make_umap_scatterplot}, \code{make_alluvial}, and \code{prettyOncoplot} for visualization.
}
}
\examples{
DLBCLone_summarize_model("Full_geneset_unweighted", optimized_model)

}
