% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flatten_feature}
\alias{flatten_feature}
\title{Flatten feature}
\usage{
flatten_feature(new_name, features_to_flatten, incoming_data)
}
\arguments{
\item{new_name}{String of the new feature name.}

\item{features_to_flatten}{Vector of features (row names) to be flattened.}

\item{incoming_data}{Matrix or data frame of features.}
}
\value{
matrix
}
\description{
Return matrix (features in rows and samples in columns) where several features are
squished (flattened) into a single one. The number and order of features to be flattened
does not matter. New row with the new name specified by user will be created, and will be positive
(value of 1) if any of the features to flatten are positive.
}
