% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handle_genome_build}
\alias{handle_genome_build}
\title{Harmonize different flavors of genome builds.}
\usage{
handle_genome_build(incoming_genome_build)
}
\arguments{
\item{incoming_genome_build}{The string specifying the genome build that is about to be harmonized.}
}
\value{
string
}
\description{
Will process different genome build flavors and return it in consistent formatting used throughout this package.
}
