% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lacy_features}
\alias{lacy_features}
\title{Features for DLBCL grouping by Lacy method.}
\format{
\subsection{\code{lacy_features}}{

A named list.
}
}
\usage{
lacy_features
}
\description{
A named list of data frames and lists needed to assemble matrix and classify
DLBCL tumors according to Lacy method described in PMID: 32187361.
}
\keyword{datasets}
