% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphgenerator_features}
\alias{lymphgenerator_features}
\title{Features for DLBCL grouping by unified method.}
\format{
\subsection{\code{lymphgenerator_features}}{

A named list.
}
}
\usage{
lymphgenerator_features
}
\description{
A named list of data frames and lists needed to assemble matrix and classify
DLBCL tumors according to the unified classification.
}
\keyword{datasets}
