% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{make_and_annotate_umap}
\alias{make_and_annotate_umap}
\title{Run UMAP and attach result to metadata}
\usage{
make_and_annotate_umap(
  df,
  metadata,
  umap_out,
  n_neighbors = 55,
  min_dist = 0,
  metric = "cosine",
  n_epochs = 1500,
  init = "spca",
  ret_model = TRUE,
  na_vals = "drop",
  join_column = "sample_id",
  seed = 12345,
  target_column,
  target_metric = "euclidean",
  target_weight = 0.5
)
}
\arguments{
\item{df}{Feature matrix with one row per sample and one column per mutation}

\item{metadata}{Metadata data frame with one row per sample and a column sample_id that
matches the row names of df. This data frame will be joined to the UMAP output.}

\item{umap_out}{Optional UMAP output from a previous run. If provided, the function
will use this model to project the data instead of re-running UMAP. This is useful
for reproducibility and for using the same UMAP model on different datasets.}

\item{n_neighbors}{Passed to UMAP2. The number of neighbors to consider when calculating the UMAP embedding.}

\item{min_dist}{Passed to UMAP2. The minimum distance between points in the UMAP embedding.}

\item{metric}{Passed to UMAP2. The distance metric to use for calculating distances between points.}

\item{n_epochs}{Passed to UMAP2. The number of epochs to run the UMAP algorithm.}

\item{init}{Passed to UMAP2. The initialization method for the UMAP algorithm.}

\item{ret_model}{additional argument}

\item{na_vals}{How to deal with NA values. Two options are "drop", which
will remove all columns containing at least one NA or "to_zero", which sets
all NA to zero and leaves the column intact.}

\item{join_column}{The column name in the metadata data frame that contains the sample IDs (default sample_id).}

\item{seed}{Passed to UMAP2. The random seed for reproducibility.}
}
\description{
Run UMAP and attach result to metadata
}
\examples{

umap_outs = make_and_annotate_umap(df=gambl_train_lymphgen,
                           min_dist = 0,
                           n_neighbors = 55,
                           init="spectral",
                           n_epochs = 1500,
                           #seed=best_params$seed,
                           metadata=gambl_train_meta_dlbclass,
                           ret_model=T,
                           metric="cosine")


}
