% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{optimize_outgroup}
\alias{optimize_outgroup}
\title{Optimize the threshold for classifying samples as "Other"}
\usage{
optimize_outgroup(
  metadata,
  predicted_labels,
  true_labels,
  other_score,
  optimize_for_dlbclass = FALSE,
  all_classes = c("MCD", "EZB", "BN2", "N1", "ST2", "Other"),
  maximize = "balanced_accuracy",
  exclude_other_for_accuracy = FALSE
)
}
\arguments{
\item{metadata}{Metadata data frame with one row per sample and a column sample_id that
matches the row names of df. This data frame will be joined to the UMAP output.}

\item{predicted_labels}{Vector of predicted labels for the samples}

\item{true_labels}{Vector of true labels for the samples}

\item{other_score}{Vector of scores for the "Other" class for each sample ()}

\item{optimize_for_dlbclass}{Set to TRUE to optimize the optimized other thresholds into
assigning lymphgen labels only if scoring a high confidence assoicated dlbclass label. Default:
FALSE}

\item{all_classes}{Vector of classes to use for training and testing.
Default: c("MCD","EZB","BN2","N1","ST2","Other")}

\item{maximize}{Metric to use for optimization.
Either "accuracy" (actual accuracy across all samples) or "balanced_accuracy"
(the mean of the balanced accuracy values across all classes).
Default: "balanced_accuracy"}

\item{exclude_other_for_accuracy}{Set to TRUE to exclude the
"Other" class from the 'lymphgen' column when calculating accuracy metrics
(passed to DLBCLone_optimize_params). Default: FALSE}
}
\value{
a list of data frames with the predictions and the UMAP input
}
\description{
Performs a post-hoc evaluation of the classification of a sample as one of
the main classes vs the outgroup/unclassified label "Other" and returns the
optimal threshold for classifying a sample as "Other" based on the ground
truth provided in the true_labels vector. It evaluates the performance
of the classifier using a range of thresholds and returns the best threshold
based on the specified metric (balanced accuracy or accuracy).
This function is not generally meant to be called directly but rather is
a helper function used by DLBCLone_optimize_params.
}
