% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{posthoc_feature_enrichment}
\alias{posthoc_feature_enrichment}
\title{Determine feature enrichment per class using truth or predicted labels}
\usage{
posthoc_feature_enrichment(
  sample_metadata,
  features,
  label_column = "lymphgen",
  truth_classes = c("BN2", "EZB", "MCD", "ST2", "N1"),
  method = "frequency",
  num_feats = 10,
  p_threshold = 0.01,
  title = NULL,
  base_size = 7,
  separate_plot_per_group = TRUE
)
}
\arguments{
\item{sample_metadata}{Data frame containing sample metadata with class labels,
by default in a column named "lymphgen". Use \code{label_column} to specify a different column.}

\item{label_column}{Name of the column containing the
class labels. The default is to use "lymphgen", the default truth class.}

\item{truth_classes}{Vector of class labels to consider (default: c("BN2","EZB","MCD","ST2","N1")).}

\item{method}{Method to determine top features: "frequency" for most abundant
features, "chi_square" for top differentially mutated features in the classes
vs all other classes (default : "frequency").}

\item{num_feats}{Number of top features to display per subtype (default: 10).}

\item{p_threshold}{Maximum P value to retain (when method is fisher)}

\item{title}{Title for the plot (default: NULL).}

\item{base_size}{Base font size used (passed to theme_Morons)}

\item{separate_plot_per_group}{If TRUE, creates separate plots for each group
and also combines them with ggarrange (default: FALSE).}
}
\value{
A ggplot2 object representing the stacked bar plot.
}
\description{
This function identifies the top N features (genes) for
each subtype based on their prevalence in the dataset using either
the truth labels or the predicted subgroups from DLBCLone.
}
\examples{
\dontrun{
library(GAMBLR.predict)

# Assuming my_DLBCLone_opt is the output from DLBCLone_optimize_params
plot_list <- posthoc_feature_enrichment(
    my_DLBCLone_opt$predictions,
    features=DLBCLone_model$features,
    method = "chi_square",
    num_feats = 10,
    title = "LymphGen"
) 
print(plot_list$bar_plot)

plot_list <- posthoc_feature_enrichment(
   label_column = "lymphgen",
   sample_metadata = my_DLBCLone_opt$predictions,
   features = my_DLBCLone_opt$features,
   method = "fisher",
   num_feats = 10,
   base_size=9
)
print(plot_list$forest_plot)
}

}
