% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample}
\alias{predict_single_sample}
\title{Predict class for a single sample without using umap_transform}
\usage{
predict_single_sample(
  test_df,
  train_df,
  train_metadata,
  best_params,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2"),
  drop_unlabeled_from_training = TRUE,
  make_plot = TRUE
)
}
\arguments{
\item{test_df}{Data frame containing the mutation status of the test sample}

\item{train_df}{Data frame containing the mutation status of the training samples}

\item{train_metadata}{Metadata for training samples with truth labels in lymphgen column}

\item{best_params}{Data frame from DLBCLone_optimize_params with the best parameters}

\item{truth_classes}{Vector of classes to use for training and testing. Default: c("EZB","MCD","ST2","N1","BN2")}

\item{drop_unlabeled_from_training}{Set to TRUE to drop unlabeled samples from the training data}

\item{make_plot}{Set to TRUE to plot the UMAP projection and predictions}
}
\value{
a list of data frames with the predictions and the UMAP input
}
\description{
Predict class for a single sample without using umap_transform
}
