% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample_DLBCLone}
\alias{predict_single_sample_DLBCLone}
\title{Predict class for a single sample without using umap_transform and plot result of classification}
\usage{
predict_single_sample_DLBCLone(
  test_df,
  train_metadata,
  optimized_model = NULL,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2", "Other"),
  seed = 12345,
  max_neighbors = 500
)
}
\arguments{
\item{test_df}{Data frame containing the mutation status of the test sample}

\item{train_metadata}{Metadata for training samples with truth labels in lymphgen column}

\item{optimized_model}{list of parameters from DLBCLone_optimize_params, neccessary UMAP output from a
previous, Data frame with the best parameters. useful for reproducibility.}

\item{truth_classes}{Vector of classes to use for training and testing. Default: c("EZB","MCD","ST2","N1","BN2")}

\item{seed}{Random seed for reproducibility}

\item{max_neighbors}{Maximum number of neighbors to consider for each sample. Default 500.}
}
\value{
a list of data frames with the predictions, the UMAP input, the model, and a ggplot object
}
\description{
Predict class for a single sample without using umap_transform and plot result of classification
}
\examples{
predict_single <- predict_single_sample_DLBCLone(
  test_df = optimize_params$features[1,],
  train_metadata = dlbcl_meta, 
  optimized_model = optimize_params_T
)

}
