% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFmodel_BL}
\alias{RFmodel_BL}
\title{BL Classifier model.}
\format{
\subsection{\code{RFmodel_BL}}{

A randomForest model.
}
}
\usage{
RFmodel_BL
}
\description{
A random forest model developed and described in PMID: 36201743. Used to
classify BL tumours into genetic subgroups.
}
\keyword{datasets}
