% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFmodel_FL}
\alias{RFmodel_FL}
\title{FL Classifier model.}
\format{
\subsection{\code{RFmodel_FL}}{

A randomForest model.
}
}
\usage{
RFmodel_FL
}
\description{
A random forest model developed and described in PMID: 37084389. Used to
classify FL tumours into genetic subgroups.
}
\keyword{datasets}
