% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFmodel_Lacy}
\alias{RFmodel_Lacy}
\title{DLBCL Classifier model.}
\format{
\subsection{\code{RFmodel_Lacy}}{

A randomForest model.
}
}
\usage{
RFmodel_Lacy
}
\description{
A random forest model developed to reproduce the DLBCL grouping approach
described in PMID: 32187361. Used to classify DLBCL tumours into HMRN groups.
}
\keyword{datasets}
