% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stacked_bar_plot}
\alias{stacked_bar_plot}
\title{Create a stacked bar plot of top features per subtype}
\usage{
stacked_bar_plot(
  DLBCLone_model,
  truth_column = "lymphgen",
  truth_classes = c("BN2", "EZB", "MCD", "ST2"),
  method = "common",
  num_feats = 10,
  title = NULL
)
}
\arguments{
\item{DLBCLone_model}{A DLBCLone model object, which can be the output of \code{DLBCLone_optimize_params}, or \code{DLBCLone_KNN}}

\item{truth_column}{Name of the column containing the true class labels (default: "lymphgen").}

\item{truth_classes}{Vector of class labels to consider (default: c("BN2","EZB","MCD","ST2")).}

\item{method}{Method to determine top features: "common" for most common features, "chi_square" for subtype vs rest significance (default : "common").}

\item{num_feats}{Number of top features to display per subtype (default: 10).}

\item{title}{Title for the plot (default: NULL).}
}
\value{
A ggplot2 object representing the stacked bar plot.
}
\description{
This function generates a stacked bar plot showing the top features (genes) for each subtype based on their prevalence in the dataset.
}
\examples{
\dontrun{
library(GAMBLR.predict)

stacked_bar_plot(
 optimize_params,
 method = "chi_square",
 num_feats = 10,
 title = "LymphGen"
) 
}

}
