% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{classify_bl}
\alias{classify_bl}
\title{Classify BL samples into genetic subgroups.}
\usage{
classify_bl(
  these_samples_metadata,
  maf_data,
  this_seq_type = "genome",
  projection = "grch37",
  output = "both",
  ashm_cutoff = 3
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{this_seq_type}{The seq_type of the samples. Only really used to retrerive mutations when maf data is not provided and to be retreived through GAMBLR.data. Defaults to genome.}

\item{projection}{The projection of the samples. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to both.}

\item{ashm_cutoff}{Numeric value indicating number of mutations for binarizing aSHM feature. Recommended to use the default value (3).}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Use the random forest prediction model to assemble the binary matrix and use it to classify BL tummors into genetic subgroups
}
\examples{
test_meta <- get_gambl_metadata(case_set = "BL-DLBCL-manuscript-HTMCP")
predictions <- classify_bl(these_samples_metadata = test_meta)
predictions <- classify_bl(these_samples_metadata = test_meta, output = "predictions")

}
