% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{DLBCLone_KNN}
\alias{DLBCLone_KNN}
\title{Run DLBCLone KNN Classification}
\usage{
DLBCLone_KNN(
  df,
  metadata,
  core_features = NULL,
  core_feature_multiplier = 1.5,
  min_k = 5,
  max_k = 60,
  truth_column = "lymphgen",
  truth_classes = c("EZB", "BN2", "ST2", "MCD", "N1", "Other"),
  predict_unlabeled = FALSE,
  plot_samples = NULL,
  DLBCLone_KNN_out = NULL
)
}
\description{
Run DLBCLone KNN Classification
}
