% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{DLBCLone_load_optimized}
\alias{DLBCLone_load_optimized}
\title{load previously saved DLBCLone model and parameters}
\usage{
DLBCLone_load_optimized(path = ".", name_prefix = "DLBCLone")
}
\arguments{
\item{path}{Path to open saved the files}

\item{name_prefix}{Prefix of the saved files, all files will be in path and start with name_prefix}
}
\value{
a reconstructed model for use with DLBCLone_predict
}
\description{
load previously saved DLBCLone model and parameters
}
\examples{
#Restore a model for use in a fresh R session
DLBCLone_myfeats_opt <- DLBCLone_load_optimized(
  path="./",
  name_prefix="DLBCLone_myfeats"
)

}
