% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{DLBCLone_load_optimized}
\alias{DLBCLone_load_optimized}
\title{load previously saved DLBCLone model and parameters}
\usage{
DLBCLone_load_optimized(path = "models/", name_prefix = "test")
}
\arguments{
\item{path}{Path to open saved the files}

\item{name_prefix}{Prefix of the saved files, all files will be in path and start with name_prefix}
}
\value{
saves the files to the specified path
}
\description{
load previously saved DLBCLone model and parameters
}
\examples{

\dontrun{
library(GAMBLR.predict)

load_optimized <- DLBCLone_load_optimized(
  path="/save_optimized/trial_folder",
  name_prefix="test_A"
)
}

}
