% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{DLBCLone_optimize_params}
\alias{DLBCLone_optimize_params}
\title{Optimize parameters for classifying samples using UMAP and k-nearest neighbor}
\usage{
DLBCLone_optimize_params(
  combined_mutation_status_df,
  metadata_df,
  umap_out,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2", "Other"),
  eval_group = "Lacy",
  min_k = 3,
  max_k = 30,
  verbose = FALSE,
  seed = 12345
)
}
\arguments{
\item{combined_mutation_status_df}{}

\item{metadata_df}{}

\item{truth_classes}{}

\item{eval_group}{}
}
\value{
List of data frames with the results of the parameter optimization
including the best model, the associated knn parameters and the annotated UMAP output
}
\description{
Optimize parameters for classifying samples using UMAP and k-nearest neighbor
}
\examples{

lymphgen_A53_DLBCLone =  DLBCLone_optimize_params( 
   lgen_feat_status, #our binary feature matrix
   a53_meta, #our metadata 
   umap_out = lymphgen_A53_all_feat_gambl, # force use existing UMAP fit
   eval_group = NULL, # use all samples for evaluating accuracy
   truth_classes = c("MCD","EZB","BN2","ST2","N1","A53","Other"))
}
