% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{DLBCLone_save_optimized}
\alias{DLBCLone_save_optimized}
\title{model storage for DLBCLone outputs}
\usage{
DLBCLone_save_optimized(
  DLBCLone_model,
  base_path = "./",
  name_prefix = "DLBCLone"
)
}
\arguments{
\item{DLBCLone_model}{The model object to save}

\item{name_prefix}{Prefix for the saved files, all files will be in path and start with name_prefix}

\item{path}{Path to save the files}
}
\value{
saves the files to the specified path
}
\description{
model storage for DLBCLone outputs
}
\examples{
DLBCLone_save_optimized(
 optimzied_params = optimized_params,
 path="/save_optimized/trial_folder",
 name_prefix="test_A"
)
}
