% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{DLBCLone_save_optimized}
\alias{DLBCLone_save_optimized}
\title{model storage for DLBCLone outputs}
\usage{
DLBCLone_save_optimized(
  optimized_params = NULL,
  path = "models/",
  name_prefix = "test"
)
}
\arguments{
\item{optimized_params}{List containing the optimized parameters from DLBCLone_optimize_params}

\item{path}{Path to save the files}

\item{name_prefix}{Prefix for the saved files, all files will be in path and start with name_prefix}
}
\value{
saves the files to the specified path
}
\description{
model storage for DLBCLone outputs
}
\examples{

\dontrun{
library(GAMBLR.predict)

DLBCLone_save_optimized(
 optimzied_params = optimized_params,
 path="/save_optimized/trial_folder",
 name_prefix="test_A"
)
}

}
