% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{make_umap_scatterplot}
\alias{make_umap_scatterplot}
\title{Make UMAP scatterplot}
\usage{
make_umap_scatterplot(
  df,
  title = NULL,
  drop_composite = TRUE,
  colour_by = "lymphgen",
  drop_other = FALSE,
  high_confidence = FALSE,
  custom_colours,
  add_labels = FALSE,
  facet = FALSE
)
}
\arguments{
\item{df}{}

\item{drop_composite}{}

\item{colour_by}{}

\item{drop_other}{}

\item{high_confidence}{}

\item{custom_colours}{}

\item{add_labels}{}

\item{facet}{If TRUE: truth, predicted, predicted_optimized, ggmarginal not used in order to fit nicely}
}
\description{
Make UMAP scatterplot
}
