% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_umap_scatterplot}
\alias{make_umap_scatterplot}
\title{Make UMAP scatterplot}
\usage{
make_umap_scatterplot(
  df,
  truth_column = "lymphgen",
  drop_composite = TRUE,
  drop_other = FALSE,
  high_confidence = FALSE,
  custom_colours,
  add_labels = FALSE,
  title = NULL
)
}
\arguments{
\item{df}{Data frame containing the UMAP coordinates and annotations.}

\item{truth_column}{Name of the column containing the labels to be plotted (default: "lymphgen").}

\item{drop_composite}{If TRUE: removes composite labels from the lymphgen column.}

\item{drop_other}{If TRUE: removes "Other" and "NOS" labels from the lymphgen column.}

\item{high_confidence}{If TRUE: filters the data to include only samples with confidence > 0.7.}

\item{custom_colours}{Custom color palette for the plot. If not provided, uses default GAMBL colors.}

\item{add_labels}{If TRUE: adds labels to the points based on the median coordinates of each group.}

\item{title}{Title for the plot. Default: NULL.}
}
\description{
Make UMAP scatterplot
}
\examples{
\dontrun{
make_umap <- make_and_annotate_umap(
  df=all_full_status,
  metadata=dlbcl_meta
)

make_umap_scatterplot(
  make_umap$df,
  drop_other = F
)
}

}
