% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_umap_scatterplot}
\alias{make_umap_scatterplot}
\title{Make UMAP scatterplot}
\usage{
make_umap_scatterplot(
  df,
  drop_composite = TRUE,
  colour_by = "lymphgen",
  drop_other = FALSE,
  high_confidence = FALSE,
  custom_colours,
  add_labels = FALSE,
  title = NULL,
  base_size = 8
)
}
\arguments{
\item{add_labels}{}
}
\description{
Make UMAP scatterplot
}
