% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample}
\alias{predict_single_sample}
\title{Predict class for a single sample without using umap_transform}
\usage{
predict_single_sample(
  test_df,
  train_df,
  train_metadata,
  best_params,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2"),
  drop_unlabeled_from_training = TRUE,
  make_plot = TRUE
)
}
\arguments{
\item{test_df}{}

\item{train_df}{}

\item{train_metadata}{}

\item{best_params}{}
}
\description{
Predict class for a single sample without using umap_transform
}
