% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample_DLBCLone}
\alias{predict_single_sample_DLBCLone}
\title{Predict class for a single sample without using umap_transform and plot result of classification}
\usage{
predict_single_sample_DLBCLone(
  test_df,
  train_metadata,
  optimized_model = NULL,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2", "Other"),
  seed = 12345,
  max_neighbors = 500
)
}
\arguments{
\item{test_df}{Data frame containing the mutation status of the test sample}

\item{train_metadata}{Metadata for training samples with truth labels in lymphgen column}

\item{optimized_model}{list of parameters from DLBCLone_optimize_params, neccessary UMAP output from a
previous, Data frame with the best parameters. useful for reproducibility.}

\item{truth_classes}{Vector of classes to use for training and testing. Default: c("EZB","MCD","ST2","N1","BN2")}

\item{seed}{Random seed for reproducibility}

\item{max_neighbors}{Maximum number of neighbors to consider for each sample. Default 500.}
}
\value{
a list of data frames with the predictions, the UMAP input, the model, and a ggplot object
}
\description{
Predict class for a single sample without using umap_transform and plot result of classification
}
\examples{

library(GAMBLR.predict)

dlbcl_meta = readr::read_tsv(system.file("extdata/dlbcl_meta_with_dlbclass.tsv",package = "GAMBLR.predict"))

all_full_status = readr::read_tsv(system.file("extdata/all_full_status.tsv",package = "GAMBLR.predict")) \%>\%
 tibble::column_to_rownames("sample_id")

make_umap <- make_and_annotate_umap(
  df=all_full_status,
  metadata=dlbcl_meta
)

optimize_params <- DLBCLone_optimize_params(  
  all_full_status, 
  dlbcl_meta,
  umap_out = make_umap,
  truth_classes = c("MCD","EZB","BN2","N1","ST2","Other"),
  optimize_for_other = T,
  min_k=5,
  max_k=23
)

predict_single <- predict_single_sample_DLBCLone(
  test_df = optimize_params$features[1,],
  train_metadata = dlbcl_meta, 
  optimized_model = optimize_params
)

}
