% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample_DLBCLone}
\alias{predict_single_sample_DLBCLone}
\title{Predict class for a single sample without using umap_transform and plot result of classification}
\usage{
predict_single_sample_DLBCLone(
  test_df,
  train_df,
  train_metadata,
  projection,
  umap_out,
  best_params,
  optimized_model = NULL,
  other_df,
  ignore_self = FALSE,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2", "Other"),
  drop_unlabeled_from_training = TRUE,
  make_plot = TRUE,
  annotate_accuracy = FALSE,
  label_offset = 2,
  title1 = "GAMBL",
  title2 = "predicted_class_for_HighConf",
  title3 = "predicted_class_for_Other",
  seed = 12345,
  max_neighbors = 500,
  other_class = "Other"
)
}
\arguments{
\item{test_df}{Data frame containing the mutation status of the test sample}

\item{train_metadata}{Metadata for training samples with truth labels in lymphgen column}

\item{optimized_model}{list of parameters from DLBCLone_optimize_params, neccessary UMAP output from a
previous, Data frame with the best parameters. useful for reproducibility.}

\item{truth_classes}{Vector of classes to use for training and testing. Default: c("EZB","MCD","ST2","N1","BN2")}

\item{seed}{Random seed for reproducibility}

\item{max_neighbors}{Maximum number of neighbors to consider for each sample. Default 500.}
}
\value{
a list of data frames with the predictions, the UMAP input, the model, and a ggplot object
}
\description{
Predict class for a single sample without using umap_transform and plot result of classification
}
\examples{

library(GAMBLR.predict)

dlbcl_meta = readr::read_tsv(system.file("extdata/dlbcl_meta_with_dlbclass.tsv",package = "GAMBLR.predict"))

predict_single <- predict_single_sample_DLBCLone(
  test_df = optimize_params$features[1,],
  train_metadata = dlbcl_meta, 
  optimized_model = optimize_params
)

}
