% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{predict_single_sample_DLBCLone}
\alias{predict_single_sample_DLBCLone}
\title{Predict class for a single sample without using umap_transform and plot result of classification}
\usage{
predict_single_sample_DLBCLone(
  test_df,
  train_df,
  train_metadata,
  projection,
  umap_out,
  best_params,
  optimized_model = NULL,
  other_df,
  ignore_self = FALSE,
  truth_classes = c("EZB", "MCD", "ST2", "N1", "BN2", "Other"),
  drop_unlabeled_from_training = TRUE,
  make_plot = TRUE,
  annotate_accuracy = FALSE,
  label_offset = 2,
  title1 = "GAMBL",
  title2 = "predicted_class_for_HighConf",
  title3 = "predicted_class_for_Other",
  seed = 12345,
  max_neighbors = 500
)
}
\arguments{
\item{test_df}{Data frame containing the mutation status of the test sample}

\item{train_df}{Data frame containing the mutation status of the training samples}

\item{train_metadata}{Metadata for training samples with truth labels in lymphgen column}

\item{umap_out}{UMAP output from a previous run. The function will use this model to project the data, useful
for reproducibility and for using the same UMAP model on different datasets.}

\item{best_params}{Data frame from DLBCLone_optimize_params with the best parameters}

\item{other_df}{Data frame containing the predictions for samples in the "Other" class}

\item{ignore_self}{Set to TRUE to avoid considering a neighbor with the same ID
distance = 0. This is usually only relevant when re-classifying labeled
samples to estimate overall accuracy}

\item{truth_classes}{Vector of classes to use for training and testing. Default: c("EZB","MCD","ST2","N1","BN2")}

\item{drop_unlabeled_from_training}{Set to TRUE to drop unlabeled samples from the training data}

\item{make_plot}{Set to TRUE to plot the UMAP projection and predictions}

\item{annotate_accuracy}{Set to true to add labels with accuracy values}

\item{label_offset}{Length of the label offset for the accuracy labels}

\item{title1}{additional argument}

\item{title2}{additional argument}

\item{title3}{additional argument}

\item{seed}{Random seed for reproducibility}
}
\value{
a list of data frames with the predictions, the UMAP input, the model, and a ggplot object
}
\description{
Predict class for a single sample without using umap_transform and plot result of classification
}
\examples{
predict_single_sample_DLBCLone(
   seed = 1234,
   test_df = test_df,
   train_df = train_df,
   train_metadata = train_metadata,
   umap_out = umap_out,
   best_params = best_params
   predictions_df = predictions_df,
   annotate_accuracy = TRUE
)

}
