% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{weighted_knn_predict_with_conf}
\alias{weighted_knn_predict_with_conf}
\title{Weighted k-nearest neighbor with confidence estimate}
\usage{
weighted_knn_predict_with_conf(
  train_coords,
  train_labels,
  test_coords,
  k,
  epsilon = 0.1,
  conf_threshold = NULL,
  na_label = "Other",
  verbose = FALSE,
  use_weights = TRUE,
  track_neighbors = TRUE,
  separate_other = TRUE,
  max_neighbors = 500
)
}
\arguments{
\item{train_coords}{Data frame containing coordinates for samples with known
labels (training samples). One row per sample, one column per feature. Usually
this would be the two UMAP components}

\item{train_labels}{Vector of labels for training cases}

\item{test_coords}{Data frame containing coordinates for samples to be classified
with the same format as train_coords}

\item{k}{The number of neigbors to consider}

\item{epsilon}{This value is added to the distances before applying weights
when use_weights is TRUE. Default: 1e-5}

\item{conf_threshold}{Minimum confidence for classifying a sample based on
it's neighbors. Below this value the sample will be assigned na_label instead}

\item{na_label}{Class to assign all samples that are not confidently assigned.
Default: Other}

\item{verbose}{Whether to print verbose outputs to console}

\item{use_weights}{Set to FALSE for all neigbors to have equal weight when
calculating the confidence}

\item{track_neighbors}{Set to TRUE to include details for the nearest neighbors of each sample}

\item{separate_other}{Set to TRUE to treat the "Other" class separately
when calculating the confidence.}

\item{max_neighbors}{Maximum number of neighbors to consider for each sample. Default 500.}
}
\value{
data frame with labels and confidence values for rows in test_coords
}
\description{
Weighted k-nearest neighbor with confidence estimate
}
